requirements.txt
SQLAlchemy~=2.0.38
python-dotenv~=1.0.1
pydantic~=2.10.6
fastapi~=0.115.8
PyJWT~=2.10.1
passlib~=1.7.4
starlette~=0.45.3
requests~=2.32.3
alembic~=1.14.1
uvicorn~=0.34.0
psycopg2
pydantic[email]
python-multipart
kafka~=1.3.5
pytest~=8.3.4
asyncpg
greenlet
httpx~=0.28.1
aiokafka~=0.12.0